LPS-51814 Unit tests
