+++ modules/portal/portal-template-soy/test/unit/com/liferay/portal/soy/SoyManagerTest.java
@@ -16,45 +16,31 @@ package com.liferay.portal.soy;
 
 import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
 import com.liferay.portal.kernel.template.Template;
-import com.liferay.portal.kernel.template.TemplateConstants;
 import com.liferay.portal.kernel.template.TemplateException;
-import com.liferay.portal.kernel.template.TemplateManagerUtil;
-import com.liferay.portal.kernel.template.TemplateResource;
-import com.liferay.portal.kernel.template.URLTemplateResource;
-import com.liferay.portal.test.runners.LiferayIntegrationJUnitTestRunner;
-
-import java.net.URL;
 
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 /**
  * @author Bruno Basto
  */
-@RunWith(LiferayIntegrationJUnitTestRunner.class)
-public class SoyTemplateTest {
+public class SoyManagerTest {
 
 	@Before
-	public void setUp() throws TemplateException {
-		TemplateManagerUtil.init();
+	public void setUp() throws Exception {
+		_soyManagerTestHelper.setUp();
 	}
 
 	@After
 	public void tearDown() {
-		Class<?> clazz = getClass();
-
-		TemplateManagerUtil.destroy(clazz.getClassLoader());
+		_soyManagerTestHelper.tearDown();
 	}
 
 	@Test
 	public void testProcessTemplateSimple() throws Exception {
-		TemplateResource templateResource = getTemplateResource("simple.soy");
-
-		Template template = TemplateManagerUtil.getTemplate(
-			TemplateConstants.LANG_TYPE_SOY, templateResource, false);
+		Template template = _soyManagerTestHelper.getTemplate("simple.soy");
 
 		template.put("namespace", "soy.test.simple");
 
@@ -69,10 +55,7 @@ public class SoyTemplateTest {
 
 	@Test
 	public void testProcessTemplateWithContext() throws Exception {
-		TemplateResource templateResource = getTemplateResource("context.soy");
-
-		Template template = TemplateManagerUtil.getTemplate(
-			TemplateConstants.LANG_TYPE_SOY, templateResource, false);
+		Template template = _soyManagerTestHelper.getTemplate("context.soy");
 
 		template.put("namespace", "soy.test.withContext");
 
@@ -89,33 +72,12 @@ public class SoyTemplateTest {
 
 	@Test(expected = TemplateException.class)
 	public void testProcessTemplateWithoutNamespace() throws Exception {
-		TemplateResource templateResource = getTemplateResource("simple.soy");
-
-		Template template = TemplateManagerUtil.getTemplate(
-			TemplateConstants.LANG_TYPE_SOY, templateResource, false);
+		Template template = _soyManagerTestHelper.getTemplate("simple.soy");
 
 		template.processTemplate(new UnsyncStringWriter());
 	}
 
-	protected TemplateResource getTemplateResource(String name) {
-		TemplateResource templateResource = null;
-
-		String resource = _TPL_PATH.concat(name);
-
-		Class<?> clazz = getClass();
-
-		ClassLoader classLoader = clazz.getClassLoader();
-
-		URL url = classLoader.getResource(resource);
-
-		if (url != null) {
-			templateResource = new URLTemplateResource(resource, url);
-		}
-
-		return templateResource;
-	}
-
-	private static final String _TPL_PATH =
-		"com/liferay/portal/soy/dependencies/";
+	private final SoyManagerTestHelper _soyManagerTestHelper =
+		new SoyManagerTestHelper();
 
 }
