+++ modules/portal/portal-template-soy/test/unit/com/liferay/portal/soy/SoyManagerTestHelper.java
@@ -1,0 +1,74 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.soy;
+
+import com.liferay.portal.kernel.template.Template;
+import com.liferay.portal.kernel.template.TemplateResource;
+import com.liferay.portal.kernel.template.URLTemplateResource;
+
+import java.net.URL;
+
+/**
+ * @author Marcellus Tavares
+ * @author Bruno Basto
+ */
+public class SoyManagerTestHelper {
+
+	public Template getTemplate(String fileName) {
+		TemplateResource templateResource = getTemplateResource(fileName);
+
+		return _soyManager.getTemplate(templateResource, false);
+	}
+
+	public void setUp() throws Exception {
+		setUpSoyManager();
+	}
+
+	public void tearDown() {
+		_soyManager.destroy();
+	}
+
+	protected TemplateResource getTemplateResource(String name) {
+		TemplateResource templateResource = null;
+
+		String resource = _TPL_PATH.concat(name);
+
+		Class<?> clazz = getClass();
+
+		ClassLoader classLoader = clazz.getClassLoader();
+
+		URL url = classLoader.getResource(resource);
+
+		if (url != null) {
+			templateResource = new URLTemplateResource(resource, url);
+		}
+
+		return templateResource;
+	}
+
+	protected void setUpSoyManager() throws Exception {
+		_soyManager = new SoyManager();
+
+		_soyManager.setTemplateContextHelper(new SoyTemplateContextHelper());
+
+		_soyManager.init();
+	}
+
+	private static final String _TPL_PATH =
+		"com/liferay/portal/soy/dependencies/";
+
+	private SoyManager _soyManager;
+
+}
