+++ modules/portal/portal-template-soy/test/unit/com/liferay/portal/soy/dependencies/simple.soy
